refactor(gateway): extract history-build helpers from run.py (slice 13 of #54962)
